# use trapezoid method more